Composer update
Moved autoloading from src/ to framework/modules/ and application/modules
Don’t need to require so many dependencies at this time.
Explicitly load the settings.php and routes.php files.
Create settings.php and data/ files when ‘composer install’ or ‘composer update’ commands are used. (these files are not to be committed to the repo. This is not finished)